#1 [chore] entity 재설게 및 매핑
