+++ was/src/main/java/qdang/group/was/domain/match/repository/MatchRepository.java
@@ -1,0 +1,4 @@
+package qdang.group.was.domain.match.repository;
+
+public interface MatchRepository {
+}
