+++ was/src/main/java/qdang/group/was/domain/phoneAuth/domain/PhoneAuth.java
@@ -1,0 +1,37 @@
+package qdang.group.was.domain.phoneAuth.domain;
+
+
+import java.time.LocalDateTime;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@Table(name = "phone_auth")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class PhoneAuth {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String phone;
+
+    private String phoneCode;
+
+    private LocalDateTime createdAt;
+
+    private LocalDateTime updatedAt;
+
+    private LocalDateTime expiredAt;
+
+    private Boolean isDeleted;
+
+    private Boolean isPhoneAuth;
+}
