+++ was/src/main/java/qdang/group/was/domain/user/controller/UserController.java
@@ -1,0 +1,14 @@
+package qdang.group.was.domain.user.controller;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequestMapping("/api/v1/users")
+@RestController
+@RequiredArgsConstructor
+@Tag(name = "User", description = "유저 API Document")
+public class UserController {
+
+}
