+++ was/src/main/java/qdang/group/was/domain/user/domain/UserRole.java
@@ -1,0 +1,5 @@
+package qdang.group.was.domain.user.domain;
+
+public enum UserRole {
+	MEMBER, ADMIN
+}
