+++ was/src/main/java/qdang/group/was/domain/user/service/UserService.java
@@ -1,0 +1,5 @@
+package qdang.group.was.domain.user.service;
+
+
+public interface UserService {
+}
