+++ was/src/main/java/qdang/group/was/domain/user/service/UserServiceImpl.java
@@ -1,0 +1,10 @@
+package qdang.group.was.domain.user.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserServiceImpl implements UserService {
+
+}
