+++ was/src/main/java/qdang/group/was/domain/userMatchProcess/domain/UserMatchProcess.java
@@ -1,0 +1,61 @@
+package qdang.group.was.domain.userMatchProcess.domain;
+
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import qdang.group.was.domain.matchProcess.domain.MatchProcess;
+import qdang.group.was.domain.user.domain.User;
+
+@Getter
+@Entity
+@Table(name = "q_user_match_process")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class UserMatchProcess {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private Integer score;
+
+    private Integer finishCushionScore;
+
+    private Integer finishBankShotScore;
+
+    private Integer ranking;
+
+    private String status;
+
+    private Integer maxHighRun;
+
+    private Integer highRun;
+
+    private Integer deltaScore;
+
+    private String turnType;
+
+    private Boolean isMyTurn;
+
+    private Integer inningCount;
+
+    private Integer succeedInningCount;
+
+    private Integer failedInningCount;
+
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "match_process_id")
+    private MatchProcess matchProcess;
+}
