+++ was/src/main/java/qdang/group/was/global/jwt/JwtService.java
@@ -9,9 +9,7 @@ import java.nio.charset.StandardCharsets;
 import java.security.Key;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
-import java.util.Base64;
 import java.util.Date;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import qdang.group.was.global.exception.ErrorType;
 import qdang.group.was.global.exception.UnauthorizedException;
