+++ was/src/main/java/qdang/group/was/global/jwt/TokenInfo.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import qdang.group.was.user.domain.User;
+import qdang.group.was.domain.user.domain.User;
 
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
