--- was/src/main/java/qdang/group/was/match/repository/MatchRepository.java
@@ -1,4 +1,0 @@
-package qdang.group.was.match.repository;
-
-public interface MatchRepository {
-}
