--- was/src/main/java/qdang/group/was/phoneAuth/domain/PhoneAuth.java
@@ -1,19 +1,0 @@
-package qdang.group.was.phoneAuth.domain;
-
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import lombok.Getter;
-
-@Getter
-@Entity
-public class PhoneAuth {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-
-}
