--- was/src/main/java/qdang/group/was/user/controller/dto/response/LoginResponse.java
@@ -1,18 +1,0 @@
-package qdang.group.was.user.controller.dto.response;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import qdang.group.was.user.service.dto.response.TokenCollection;
-
-@Getter
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class LoginResponse {
-
-	private String accessToken;
-	private String refreshToken;
-
-	public static LoginResponse from(TokenCollection jwtToken) {
-		return new LoginResponse(jwtToken.getAccessToken(), jwtToken.getRefreshToken());
-	}
-}
