--- was/src/main/java/qdang/group/was/user/domain/UserRole.java
@@ -1,5 +1,0 @@
-package qdang.group.was.user.domain;
-
-public enum UserRole {
-	MEMBER, ADMIN
-}
