--- was/src/main/java/qdang/group/was/user/exception/ConflictUserNameException.java
@@ -1,11 +1,0 @@
-package qdang.group.was.user.exception;
-
-import qdang.group.was.global.exception.BusinessException;
-import qdang.group.was.global.exception.ErrorType;
-
-public class ConflictUserNameException extends BusinessException {
-
-	public ConflictUserNameException() {
-		super(ErrorType.ALREADY_EXIST_USER_EXCEPTION);
-	}
-}
