--- was/src/main/java/qdang/group/was/user/exception/NotFoundUserException.java
@@ -1,11 +1,0 @@
-package qdang.group.was.user.exception;
-
-import qdang.group.was.global.exception.BusinessException;
-import qdang.group.was.global.exception.ErrorType;
-
-public class NotFoundUserException extends BusinessException {
-
-	public NotFoundUserException() {
-		super(ErrorType.NOT_FOUND_USER_EXCEPTION);
-	}
-}
