--- was/src/main/java/qdang/group/was/user/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package qdang.group.was.user.repository;
-
-import java.util.Optional;
-import qdang.group.was.match.domain.Match;
-import qdang.group.was.user.domain.User;
-
-public interface UserRepository {
-
-	void save(User user);
-
-	Optional<User> findByUserId(String username);
-}
