--- was/src/main/java/qdang/group/was/user/service/UserService.java
@@ -1,5 +1,0 @@
-package qdang.group.was.user.service;
-
-
-public interface UserService {
-}
