--- was/src/main/java/qdang/group/was/user/service/UserServiceImpl.java
@@ -1,10 +1,0 @@
-package qdang.group.was.user.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserServiceImpl implements UserService {
-
-}
